// src/shared/models/notes.ts
export class Notes {
    title: String = '';
    body: String = '';
    createdAt: Date = new Date;
    updatedAt: Date = new Date;
    color: String = '';
    // images: imagesArr =
    labelName: String = '';
    reminder: String = '';
    // collabrators: Array

    constructor() {

    }
}

// export class imagesArr {
//     id: number = 1;
//     link: String = '';
// }